+++ resources/assets/es6/newnote.js
@@ -1,0 +1,8 @@
+//newnote.js
+
+import enableLocateButton from './newnote-button';
+import persistFormData from './persist-form';
+
+let button = document.querySelector('#locate');
+enableLocateButton(button);
+persistFormData();